Modify device_erase_size to fix CH559 code flash area
For CH559, device_erase_size was specified as 0x1d. But this is
too small to write a larger file. CH559 has 64KB flash ROM and
the first 60KB is used for application code. Following 1KB is
used for application data or code. And the last 3KB is for
boot loader and configuration. As the block size is 1KB, this
device_erase_size should be set to 60.